# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iris API",
    "description": "Self-hosted media aggregator + streamer — internal household API.",
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/metadata/tmdb/search": {
      "get": {
        "tags": [
          "metadata"
        ],
        "summary": "Typeahead suggestions for the search page. Proxies TMDB's\n`/search/multi` so the API key stays server-side. Empty array on no\nresults, missing TMDB config, or upstream failure (typeahead is best-\neffort: never block typing on a flaky network).",
        "operationId": "search_tmdb",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Typeahead suggestions (empty when TMDB unconfigured)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TmdbSuggestion"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/SortField"
                }
              ]
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/SortOrder"
                }
              ]
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/MediaKind"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated tracker results + library matches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/details": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Rich preview for a single torrent. Powers the search-result preview\ndialog. Provider-specific shape is normalised to a single\n`TorrentDetails` so web + TV consume one structure.",
        "operationId": "details",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Provider id from the search hit (`provider_id` field).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "Provider-specific opaque id from the search hit (`external_id`).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Normalised torrent detail view",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentDetails"
                }
              }
            }
          },
          "400": {
            "description": "Unknown provider"
          },
          "404": {
            "description": "Provider exposes no detail page for this id"
          }
        }
      }
    },
...
    "/api/torrents/preview": {
      "post": {
        "tags": [
          "torrents"
        ],
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Parsed torrent metadata (no ingest)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TorrentPreview"
                }
              }
            }
          },
          "400": {
            "description": "Unknown provider, magnet source, or parse error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AggregatedResults": {
        "type": "object",
        "required": [
          "results",
          "providers"
        ],
        "properties": {
          "parsed_query": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ParsedQueryInfo",
                "description": "SCENE-parsed view of the user query when the raw `q` looked\nlike a SCENE-style request (e.g. `Classroom of the Elite S04E11`).\nFrontend uses this to render a \"Showing results for X · S04E11\"\nbanner; absent when the parser saw nothing useful."
              }
            ]
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderResultMeta"
            }
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            }
          }
        }
      },
      "AudioInfo": {
        "type": "object",
        "properties": {
          "bitrate_kbps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "channels": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "codec": {
            "type": [
              "string",
              "null"
            ]
          },
          "commercial_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Atmos / Dolby commercial name when `MediaInfo` flagged it."
          },
          "default": {
            "type": "boolean"
          },
          "lang": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "DescriptionFormat": {
        "type": "string",
        "description": "Rich detail view for a single torrent — the kind of payload a tracker's\ntorrent-detail page would normally show. Powers the search-result\npreview dialog. Optional everywhere because shapes vary wildly across\nindexers; the typed fields are the lowest common denominator we feel\ncomfortable surfacing in a unified UI.\nWhat `TorrentDetails::description` is encoded in — indexers don't\nagree on a format. The frontend dispatches the right renderer\n(`BBCode` parser, sanitised HTML, raw text) off this. Defaults to\n[`DescriptionFormat::Bbcode`] when absent so older provider payloads\n(torr9 was the only source originally) keep working unchanged.",
        "enum": [
          "bbcode",
          "html",
          "plain"
        ]
      },
...
      "LibraryMatch": {
        "type": "object",
        "description": "A library item matching the search query, surfaced ABOVE tracker\nresults by the clients (\"you already have this\"). Built from the\nSCENE-normalised collection key, so a different release of the same\nwork matches — unlike the infohash-keyed `already_in_library` flag\non individual results, which deliberately only marks the exact\nrelease (see `ranking.rs`: other languages/cuts must stay grabbable).",
        "required": [
          "collection_id",
          "display_title",
          "kind",
          "is_anime",
          "torrent_count",
          "episode_count"
        ],
        "properties": {
          "collection_id": {
            "type": "string"
          },
          "display_title": {
            "type": "string"
          },
          "episode_count": {
            "type": "integer",
            "format": "int64"
          },
          "episode_file_idx": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "episode_infohash": {
            "type": [
              "string",
              "null"
            ]
          },
          "episode_number": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "episode_season": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Set when the query named one specific episode the library owns:\nthe exact owned file, ready for a direct `/watch` deep-link."
          },
          "is_anime": {
            "type": "boolean"
          },
          "kind": {
            "type": "string",
            "description": "`\"movie\"` | `\"tv\"` — same vocabulary as `MediaKind`."
          },
          "representative_infohash": {
            "type": [
              "string",
              "null"
            ],
            "description": "Fallback navigation target (most recently played torrent)."
          },
          "season_episode_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Set when the query was season-scoped (\"vikings s03\"): how many\nepisodes of that season the library holds."
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "torrent_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "MediaInfoSummary": {
        "type": "object",
        "properties": {
          "audio": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudioInfo"
            }
          },
          "subtitles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubInfo"
            }
          },
          "video": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/VideoInfo"
              }
            ]
          }
        }
      },
      "MediaKind": {
        "type": "string",
        "enum": [
          "movie",
          "tv"
        ]
      },
...
      "ParsedQueryInfo": {
        "type": "object",
        "description": "Surface of the SCENE parser run against the user's raw query string.\nLives here so iris-api's ranking module can construct it without\npulling iris-providers into iris-core.",
        "required": [
          "title"
        ],
        "properties": {
          "episode": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "season": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ProviderResultMeta": {
        "type": "object",
        "required": [
          "id",
          "current_page",
          "limit"
        ],
        "properties": {
          "current_page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "limit": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "total_pages": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ResolveBody": {
        "type": "object",
        "required": [
          "provider_id",
          "external_id"
        ],
        "properties": {
          "external_id": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          },
          "tmdb_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional TMDB id, captured at search time. We persist it on ingest so\nthe library/continue-watching endpoints can ship posters without a\nfuzzy title-year lookup. Frontends are expected to pass this through\nfrom the search hit when available."
          }
        }
      },
      "SearchResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AggregatedResults"
          },
          {
            "type": "object",
            "required": [
              "library_matches"
            ],
            "properties": {
              "library_matches": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/LibraryMatch"
                }
              }
            }
          }
        ],
        "description": "`AggregatedResults` + the library rows. `flatten` keeps the wire\nshape byte-compatible for deployed clients — `library_matches` is\npurely additive (TV ignores unknown keys, TS fields are optional)."
      },
...
      "SortField": {
        "type": "string",
        "enum": [
          "title",
          "size",
          "seeders",
          "leechers",
          "uploaded"
        ]
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
      "SubInfo": {
        "type": "object",
        "properties": {
          "default": {
            "type": "boolean"
          },
          "forced": {
            "type": "boolean"
          },
          "format": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"UTF-8\"`, `\"PGS\"`, `\"VobSub\"`, etc."
          },
          "lang": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TmdbKind": {
        "type": "string",
        "enum": [
          "movie",
          "tv"
        ]
      },
      "TmdbSuggestion": {
        "type": "object",
        "required": [
          "kind",
          "tmdb_id",
          "title"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/TmdbKind"
          },
          "overview": {
            "type": [
              "string",
              "null"
            ]
          },
          "poster_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "tmdb_id": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "year": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "TorrentDetails": {
        "type": "object",
        "required": [
          "provider_id",
          "external_id",
          "title"
        ],
        "properties": {
          "age": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-formatted age string from the provider when available\n(e.g., \"6 minutes\", \"2 hours\"). When absent the frontend can\nderive its own from `uploaded_at`."
          },
          "category": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Description body. Encoding depends on [`Self::description_format`]."
          },
          "description_format": {
            "$ref": "#/components/schemas/DescriptionFormat",
            "description": "Encoding of `description`. Always set on new payloads; absent\nvalues default to [`DescriptionFormat::Bbcode`] for backward\ncompatibility with pre-c411 clients."
          },
          "exclusive": {
            "type": "boolean"
          },
          "external_id": {
            "type": "string"
          },
          "file_count": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "file_size_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "freeleech": {
            "type": "boolean"
          },
          "leechers": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "media_info": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MediaInfoSummary",
                "description": "Server-side parsed view of the NFO so web + TV consume the same\nshape rather than each having to ship a `MediaInfo` regex tower."
              }
            ]
          },
          "nfo": {
            "type": [
              "string",
              "null"
            ],
            "description": "Raw `MediaInfo` text dump. Web exposes this in a collapsible\n`<details>` for power users; TV doesn't show it."
          },
          "provider_id": {
            "type": "string"
          },
          "seeders": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "times_completed": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "uploaded_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "uploader": {
            "type": [
              "string",
              "null"
            ]
          },
          "views": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "TorrentFilePreview": {
        "type": "object",
        "required": [
          "index",
          "path",
          "size_bytes",
          "is_video"
        ],
        "properties": {
          "extension": {
            "type": [
              "string",
              "null"
            ],
            "description": "Lowercase extension (no dot), e.g. \"mkv\", \"mp4\", \"srt\"."
          },
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "is_video": {
            "type": "boolean",
            "description": "Heuristic: `true` when the extension is a typical video container."
          },
          "path": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "TorrentPreview": {
        "type": "object",
        "required": [
          "infohash",
          "name",
          "total_size_bytes",
          "piece_length",
          "piece_count",
          "announce_urls",
          "files"
        ],
        "properties": {
          "announce_urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TorrentFilePreview"
            }
          },
          "infohash": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "piece_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "piece_length": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "total_size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "VideoInfo": {
        "type": "object",
        "properties": {
          "bitrate_kbps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "codec": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "fps": {
            "type": [
              "number",
              "null"
            ],
            "format": "float"
          },
          "hdr": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"HDR10\"`, `\"Dolby Vision\"`, `\"SDR\"` etc. Best-effort heuristic."
          },
          "resolution": {
            "type": [
              "string",
              "null"
            ],
            "description": "`\"1920x1080\"` style."
          }
        }
      },
...
    }
  },
...
}